Update devDependencies and use eslint
JIRA: https://issues.jboss.org/browse/RHMAP-5745

Motivation:
Some projects were failing to build because of invalid grunt
peerDependency when grunt released a new version and different plugins
were trying to pull in conflicting versions. All of the plugins have
been pretty good at updating to be compatible with the latest versions,
so we're updating to those versions, and updating to latest version of
grunt.

It turns out that in lots of cases, other devDependencies were out of
date also, so they're also being updated.

The version of turbo-test-runner though, seems important -- updating to
the latest leads to a non-successful run of `grunt fh-accept`.

In addition to that, to save another round-trip through all of the
projects, we're moving from jshint to eslint for linting and code style.

Modification:
* Update all devDependencies to latest versions
* Install eslint as devDependency
* Remove plato and jshint devDependencies
* Replace .jshintrc with .eslintrc.json
* Add version script to package.json for `npm version`
* Run fh-test instead of fh-coverage in default grunt task

Result:
New grunt-fh-build plugin