Replacing make with grunt instructions.
Motivation:
When trying to follow the instructions in the README.md make is used to
run the tests. I've not found any trace of a Makefile and unless I'm
missing something this project is using Grunt.

Modifications:
Replace make with grunt and updated to examples to use correct targets
for Grunt.
I also removed the instructions for plato as I did not find it as a
dependency for the project nor could I find anything in the Gruntfile
with it.

Result:
The instructions can now be followed to run the tests suites.