Deleting files when the associated submission is deleted.
Motivation:
Currently when a submission is deleted and the submission has file(s)
associated with it, the files are not deleted.

Modifications:
Added a step to extract the groupIds from the submission then remove
each of the files from the MongoDB collection.

Result:
Files are deleted when the submission is deleted.

All acceptance tests should pass:
$ grunt fh:accept

When deleting a submission in studio you can inspect the MongoDB
collection and verify that the entries in the file storage are removed:
$ mongo
> use testing_dev

Create a form in Studio with a File field and deploy it. Submit a file and verify
that it exists in MongoDB:
> db.formsubmissions.find({"formFields.fieldValues.fileName": "filename"}, {"formFields.fieldValues": 1});
Use the groupId from the output in the following query:
> db.fileStorage.files.find({'_id': ObjectId("groupId")});

Next, delete the submission and then rerun the last command, and there
should it should no longer exist in the database.

JIRA:
https://issues.jboss.org/browse/RHMAP-5448