Fixing test_getSubmission deleteSubmission.
Motivation:
While working on a task[1] that modifies deleteSubmission.js I noticed that
test_getSubmission.js was failing because the submission to be deleted
was not found (the call to findOne returned null as it's result and
there is no error).

Upon closer inspection of the test I noticed that the call to
deleteSubmission is using the TEST_SUBMISSION_FORMID which seems
incorrect to me. I would have though this should have been
TEST_SUBMISSION_ID instead and is also the assumption I made for the
related task mentioned above.

Modifications:
Just changing the call to deleteSubmission to delete the submission will
fail other tests as they depend on it to be present.
So I moved the call to deleteSubmission to the tearDown method and updated
the argument to be TEST_SUBMISSION_ID instead of
TEST_SUBMISSION_FORM_ID.

Result:
All acceptance test pass.

[1] https://issues.jboss.org/browse/RHMAP-5448